Bug 2046840 - Simplify path processing code r=firefox-svg-reviewers,dholbert
Differential Revision: https://phabricator.services.mozilla.com/D306190

UltraBlame original commit: 571b86eaa3818c9fdeddada8e8624045a08b3f80

+++ dom/svg/SVGPathData.cpp
@@ -72,7 +72,7 @@ bool SVGPathData::GetDistancesFromOriginToEndsOfVisibleSegments(
 
 already_AddRefed<dom::SVGPathSegment> SVGPathData::GetPathSegmentAtLength(
     dom::SVGPathElement* aPathElement, Span<const StylePathCommand> aPath,
-    float aDistance) {
+    double aDistance) {
   SVGPathTraversalState state;
 
   for (const auto& cmd : aPath) {
