+++ dom/svg/SVGPathData.h
@@ -2,8 +2,6 @@
 
 
 
-
-
 #ifndef DOM_SVG_SVGPATHDATA_H_
 #define DOM_SVG_SVGPATHDATA_H_
 
@@ -69,7 +67,7 @@ class SVGPathData {
 
   static already_AddRefed<dom::SVGPathSegment> GetPathSegmentAtLength(
       dom::SVGPathElement* aPathElement, Span<const StylePathCommand> aPath,
-      float aDistance);
+      double aDistance);
 
   void GetMarkerPositioningData(float aZoom, nsTArray<SVGMark>* aMarks) const;
